% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitosegm15.R
\name{fitoseg15}
\alias{fitoseg15}
\title{fitoseg15}
\usage{
fitoseg15(I)
}
\arguments{
\item{I}{La imagen de entrada, imagen RGB (rojo, verde, azul).}
}
\value{
Devuelve la imagen segmentada
}
\description{
Esta función toma una imagen como entrada y devuelve una imagen
     segmentada. La segmentación se basa en la diferencia entre los canales
     rojo y verde de la imagen original.
}
\examples{
library(EBImage)
I<-readImage("IMG_20230606_114941.jpg")
fitoseg15<- function (I)
{
R <-EBImage::channel(I, "red")
G <-EBImage::channel(I, "green")
B <-EBImage::channel(I, "blue")
x<-(0.441*R-0.811*G+0.385*B+18.78745+188.0176)/417.4351
b<-scales::rescale(matrix(x, ncol = ncol(x), nrow=nrow(x)))
o<-EBImage::otsu(b)
p<-b<o
Rx<-p*R
Gx<-p*G
Bx<-p*B
If<-EBImage::rgbImage(Rx,Gx,Bx)
return(If)
}
x<-fitoseg15(I)
display(x)
}
