% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitosegm16.R
\name{fitoseg16}
\alias{fitoseg16}
\title{fitoseg16}
\usage{
fitoseg16(I, c = 1, b = 0.2, tamaño = 31, sigma = 11)
}
\arguments{
\item{I}{La imagen de entrada, imagen RGB (rojo, verde, azul).}

\item{c}{contraste (valor predeterminado: 1)}

\item{b}{brillo (valor predeterminado: 0.2)}

\item{tamaño}{(valor predeterminado: 31): Este parámetro se relaciona con el
tamaño de máscara que se utiliza durante el procesamiento.}

\item{sigma}{(valor predeterminado: 11): Es probable que este parámetro esté
relacionado con un filtro Gaussiano que se aplica en algún punto del
proceso}
}
\value{
Devuelve la imagen segmentada
}
\description{
Esta función toma una imagen como entrada y devuelve una imagen
     segmentada. La segmentación se basa en la diferencia entre los canales
     rojo y verde de la imagen original.
}
\examples{
library(EBImage)
I<-readImage("IMG_20230606_114954.jpg")
fitoseg16<-function(I,c=1, b=.2, tamaño = 31, sigma=11)
{
f<-EBIamge::filter2(I, EBImage::makeBrush(size=999, "Gaussian", sigma=21))
r<-EBImage::channel(f,"red")
g<-EBImage::channel(f,"green")
b<-EBImage::channel(f,"blue")
x<-g>r&g>b
Rx<-x*EBImage::channel(I,"red")
Gx<-x*EBImage::channel(I,"green")
Bx<-x*EBImage::channel(I,"blue")
If<-EBImage::rgbImage(Rx,Gx,Bx)
return(If)
}
x<-fitoseg16(I)
display(x)
}
