% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitosegm17.R
\name{fitoseg17}
\alias{fitoseg17}
\title{fitoseg17}
\usage{
fitoseg17(I, c = 1, b = 0.2, tamaño = 31, sigma = 11)
}
\arguments{
\item{I}{La imagen de entrada, imagen RGB (rojo, verde, azul).}

\item{c}{contraste (valor predeterminado: 1)}

\item{b}{brillo (valor predeterminado: 0.2)}

\item{tamaño}{(valor predeterminado: 31): Este parámetro se relaciona con el
tamaño de máscara que se utiliza durante el procesamiento.}

\item{sigma}{(valor predeterminado: 11): Es probable que este parámetro esté
relacionado con un filtro Gaussiano que se aplica en algún punto del
proceso}
}
\value{
Devuelve la imagen segmentada
}
\description{
Esta función toma una imagen como entrada y devuelve una imagen
     segmentada. La segmentación se basa en la diferencia entre los canales
     rojo y verde de la imagen original.
}
\examples{
library(EBImage)
library(scales)
I<-readImage("IMG_20230606_114941.jpg")
fitoseg17<-function(I,c=1, b=.2, tamaño = 31, sigma=11)
{
f<-EBImage::filter2(I, EBImage::makeBrush(size=999, "Gaussian", sigma=21))
r<-EBImage::channel(f,"red")
g<-EBImage::channel(f,"green")
b<-EBImage::channel(f,"blue")
c<-scales::rescale(matrix(f, ncol = ncol(f), nrow=nrow(f)))
o<-EBImage::otsu(c)
p<-c>o
Rx<-p*EBImage::channel(I,"red")
Gx<-p*EBImage::channel(I,"green")
Bx<-p*EBImage::channel(I,"blue")
If<-EBImage::rgbImage(Rx,Gx,Bx)
return(If)
}
x<-fitoseg17(I)
display(x)
}
