% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitosegm2.R
\name{fitoseg2}
\alias{fitoseg2}
\title{fitoseg2}
\usage{
fitoseg2(I)
}
\arguments{
\item{I}{La imagen de entrada, imagen RGB (rojo, verde, azul).}
}
\value{
Devuelve la imagen segmentada
}
\description{
Esta función toma una imagen como entrada y devuelve una imagen
     segmentada. La segmentación se basa en la diferencia entre los canales
     rojo y verde de la imagen original.
}
\examples{
library(EBImage)
I<-readImage("IMG_20230606_114757.jpg")
fitoseg2<-function(I)
{
R <-EBImage::channel(I, "red")
G <-EBImage::channel(I, "green")
B <-EBImage::channel(I, "blue")
x<-2.5*(B-G)/(B+6*G-7.5*R+1)
b<-scales::rescale(matrix(x, ncol = ncol(x), nrow=nrow(x)))
o<-EBImage::otsu(b)
p<-b<o
Rx<-p*R
Gx<-p*G
Bx<-p*B
If<-EBImage::rgbImage(Rx,Gx,Bx)
return(If)
}
x<-fitoseg2(x)
display(x)
}
